(feat) Build GameLobby, move functionality to Game
I got GameLobby to work, but then noticed that to reduce code duplication, I needed Game state to be kept in GameLobby instead Game, and then passed to Game as props.  I opted instead, to add the GameLobby functionality to Game and comment out GameLobby.
I did not delete GameLobby, in case others think it should be a separate component.

was needed to use Player component.